+++ src/commands/expressions/operands/index.ts
@@ -4,7 +4,13 @@ export {
   operand,
   type RawOperand,
 } from "./operand.js";
-export { type ImplicitPath, Path, path, type RawPath } from "./path.js";
+export {
+  type ImplicitPath,
+  literal,
+  Path,
+  path,
+  type RawPath,
+} from "./path.js";
 export {
   type ImplicitValue,
   type RawValue,